fix user settings view

+++ openistorm/settings/models.py
@@ -6,8 +6,8 @@ from django.contrib.gis.geos import Point
 class Setting(models.Model):
     # user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.DO_NOTHING, related_name='settings', primary_key=True)
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.DO_NOTHING, related_name='settings')
-    sl_reference = models.FloatField(blank=True, null=True)
-    sl_notification_threshold = models.FloatField(blank=True, null=True)
+    sl_reference = models.IntegerField(blank=True, null=True)
+    sl_notification_threshold = models.IntegerField(blank=True, null=True)
 
     def __unicode__(self):
         return str(self.user.email)
